+++ src/legacy/cxgo_client__gui__guirank.go
@@ -218,16 +218,16 @@ func nox_xxx_guiDrawRank_46E870() int32 {
 	v30 = v46 + 1
 	v45[1] = 0
 	v45[2] = 0
-	dword_5d4594_1090100 = unsafe.Pointer(nox_gui_newStaticText_489300(dword_5d4594_1090048, 1088, 0, v46, dword_5d4594_1090040, v46+1, (*gui.WindowData)(unsafe.Pointer(&v49[0])), (*nox_staticText_data)(unsafe.Pointer(&v45[0]))))
+	dword_5d4594_1090100 = unsafe.Pointer(nox_gui_newStaticText_489300(dword_5d4594_1090048, 1088, 0, v46, int32(dword_5d4594_1090040), v46+1, (*gui.WindowData)(unsafe.Pointer(&v49[0])), (*nox_staticText_data)(unsafe.Pointer(&v45[0]))))
 	*(*uint32)(unsafe.Pointer(&v49[68])) = nox_color_white_2523948
 	v45[0] = int32(uintptr(unsafe.Pointer(nox_strman_loadString_40F1D0(internCStr("WindowDir:Empty"), nil, internCStr("C:\\NoxPost\\src\\client\\Gui\\guirank.c"), 1782))))
-	dword_5d4594_1090112 = uint32(uintptr(unsafe.Pointer(nox_gui_newStaticText_489300(dword_5d4594_1090048, 1088, 0, v20, dword_5d4594_1090040, v30, (*gui.WindowData)(unsafe.Pointer(&v49[0])), (*nox_staticText_data)(unsafe.Pointer(&v45[0]))))))
+	dword_5d4594_1090112 = uint32(uintptr(unsafe.Pointer(nox_gui_newStaticText_489300(dword_5d4594_1090048, 1088, 0, v20, int32(dword_5d4594_1090040), v30, (*gui.WindowData)(unsafe.Pointer(&v49[0])), (*nox_staticText_data)(unsafe.Pointer(&v45[0]))))))
 	*(*uint32)(unsafe.Pointer(&v49[68])) = *memmap.PtrUint32(0x85B3FC, 940)
 	v45[0] = int32(uintptr(unsafe.Pointer(nox_strman_loadString_40F1D0(internCStr("WindowDir:Empty"), nil, internCStr("C:\\NoxPost\\src\\client\\Gui\\guirank.c"), 1790))))
-	dword_5d4594_1090108 = uint32(uintptr(unsafe.Pointer(nox_gui_newStaticText_489300(dword_5d4594_1090048, 1088, 0, v47, dword_5d4594_1090040, v30, (*gui.WindowData)(unsafe.Pointer(&v49[0])), (*nox_staticText_data)(unsafe.Pointer(&v45[0]))))))
+	dword_5d4594_1090108 = uint32(uintptr(unsafe.Pointer(nox_gui_newStaticText_489300(dword_5d4594_1090048, 1088, 0, v47, int32(dword_5d4594_1090040), v30, (*gui.WindowData)(unsafe.Pointer(&v49[0])), (*nox_staticText_data)(unsafe.Pointer(&v45[0]))))))
 	*(*uint32)(unsafe.Pointer(&v49[68])) = *memmap.PtrUint32(0x8531A0, 2572)
 	v45[0] = int32(uintptr(unsafe.Pointer(nox_strman_loadString_40F1D0(internCStr("TeamPlayerRank"), nil, internCStr("C:\\NoxPost\\src\\client\\Gui\\guirank.c"), 1798))))
-	*memmap.PtrPtr(0x5D4594, 1090104) = unsafe.Pointer(nox_gui_newStaticText_489300(dword_5d4594_1090048, 1088, 0, 0, dword_5d4594_1090040, v30, (*gui.WindowData)(unsafe.Pointer(&v49[0])), (*nox_staticText_data)(unsafe.Pointer(&v45[0]))))
+	*memmap.PtrPtr(0x5D4594, 1090104) = unsafe.Pointer(nox_gui_newStaticText_489300(dword_5d4594_1090048, 1088, 0, 0, int32(dword_5d4594_1090040), v30, (*gui.WindowData)(unsafe.Pointer(&v49[0])), (*nox_staticText_data)(unsafe.Pointer(&v45[0]))))
 	result = int32(uintptr(unsafe.Pointer(dword_5d4594_1090048)))
 	dword_587000_145664 = 1
 	return result
